Complete '01-cascade-fix' exercise
Edited the style.css file to practice the cascade. Cascading gives the precedance to the style that appears later in the stylesheet. But for some styles like id having highest precedance then the chained classes then the acual element style and last the *. The more specific the style is more precedance it gets.