Complete '01-flex-center' exercise
Learned how to use 'display: flex' to use flexbox or initialize it.
Learned to use 'justify-content' to align div at center on start line which is where header usually is.
Learned to use 'align items' to align all the items in parent container.
Used all the above flex properties in parent container.